Move unit validator into reusable utils module.
